requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
psycopg2-binary==2.9.11
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dotenv==1.2.1
qt-material==2.17
requests==2.32.5
shiboken6==6.10.0
urllib3==2.6.2
